Backend/requirements.txt
fastapi
uvicorn
requests
yfinance
python-dotenv
supabase
pydantic
